split normalization code into separate file
